Creating an anonymous chat on Python without identification and registration
Online communication is a great opportunity in real time, which includes many opportunities to get to know people and the world around you. However, this is not always safe. I suggest trying a chat, where there is no need to identify your data, but there is an opportunity to get to know the interlocutor through communication alone.